added conditional step

+++ Jenkinsfile
@@ -87,15 +87,11 @@ pipeline {
                         env.containerID="${bat(script: 'docker ps -q -f name=c-shivam01-master', returnStdout: true).trim()}"
                         echo "Running containerID is "
                         echo env.containerID
-                    }
-                    when{
-                        expression{
-                            return env.containerID!=null
-                        }
-                    }
-                    steps{
-                        echo "Stop container and remove from stopped container list too"
-                        bat "docker stop env.containerID && docker rm env.containerID"
+                        
+                        if(env.containerID!=null){
+                            echo "Stop container and remove from stopped container list too"
+                            bat "docker stop env.containerID && docker rm env.containerID"
+                        }   
                     }
                 }
                 stage("Publish Docker Image to DockerHub"){
